[UserCenter] about-us: extract component Label
