+++ __tests__/containers/user-center/about-us/aboutUs.test.js
@@ -1,4 +1,3 @@
-// import 'react-native';
 import React from 'react';
 import { Linking } from 'react-native';
 import renderer from 'react-test-renderer';
@@ -6,7 +5,7 @@ import { shallow } from 'enzyme';
 import * as chai from 'chai';
 import ListItem from 'react-native-elements/src/list/ListItem';
 
-import AboutUs, { Label } from '../../../../src/containers/user-center/about-us/AboutUs';
+import AboutUs from '../../../../src/containers/user-center/about-us/AboutUs';
 
 describe('test user center / about us', () => {
   it('renders correctly', () => {
@@ -36,14 +35,5 @@ describe('test user center / about us', () => {
     chai.expect(wrapper.find({ title: '开源 - GitHub' })).to.have.length(1);
     chai.expect(wrapper.find({ title: 'QQ 交流群: 529600495' })).to.have.length(1);
     chai.expect(wrapper.find({ title: '论坛: https://forum.growth.ren' })).to.have.length(1);
-  });
-
-  it('test for rendering Label', () => {
-    const tree = renderer.create(
-      <Label text="LabelText" />,
-    );
-
-    const treeJson = tree.toJSON();
-    expect(treeJson).toMatchSnapshot();
   });
 });
